Factor records fixtures into seperate conftest
