+++ db/tests/records/test_filtering.py
@@ -2,30 +2,9 @@ import re
 import pytest
 from datetime import datetime
 
-from sqlalchemy import MetaData, Table
 from sqlalchemy_filters.exceptions import BadFilterFormat, FilterFieldNotFound
 
 from db import records
-
-
-ROSTER = "Roster"
-FILTER_SORT = "filter_sort"
-
-
-@pytest.fixture
-def roster_table_obj(engine_with_roster):
-    engine, schema = engine_with_roster
-    metadata = MetaData(bind=engine)
-    roster = Table(ROSTER, metadata, schema=schema, autoload_with=engine)
-    return roster, engine
-
-
-@pytest.fixture
-def filter_sort_table_obj(engine_with_filter_sort):
-    engine, schema = engine_with_filter_sort
-    metadata = MetaData(bind=engine)
-    roster = Table(FILTER_SORT, metadata, schema=schema, autoload_with=engine)
-    return roster, engine
 
 
 def test_get_records_filters_using_col_str_names(roster_table_obj):
