+++ db/tests/records/test_records.py
@@ -1,25 +1,5 @@
 import pytest
-from sqlalchemy import MetaData, Table
 from db import records
-
-ROSTER = "Roster"
-FILTER_SORT = "filter_sort"
-
-
-@pytest.fixture
-def roster_table_obj(engine_with_roster):
-    engine, schema = engine_with_roster
-    metadata = MetaData(bind=engine)
-    roster = Table(ROSTER, metadata, schema=schema, autoload_with=engine)
-    return roster, engine
-
-
-@pytest.fixture
-def filter_sort_table_obj(engine_with_filter_sort):
-    engine, schema = engine_with_filter_sort
-    metadata = MetaData(bind=engine)
-    roster = Table(FILTER_SORT, metadata, schema=schema, autoload_with=engine)
-    return roster, engine
 
 
 def test_get_records_gets_all_records(roster_table_obj):
